move testing payload generate to scripts
